socketjoy/requirements.txt
bidict==0.21.2
dnspython==1.16.0
evdev==1.3.0
eventlet==0.30.0
greenlet==0.4.17
python-engineio==4.0.0
python-socketio==5.0.4
qrcode==6.1
six==1.15.0
